+++ Source/SwiftLintFramework/Rules/Lint/ExpiringTodoRule.swift
@@ -34,7 +34,9 @@ public struct ExpiringTodoRule: ConfigurationProviderRule, OptInRule {
         ],
         triggeringExamples: [
             Example("// TODO: [10/14/2019]\n"),
-            Example("// FIXME: [10/14/2019]\n")
+            Example("// FIXME: [10/14/2019]\n"),
+            Example("// FIXME: [1/14/2019]\n"),
+            Example("// FIXME: [10/4/2019]\n")
         ]
     )
 
@@ -44,7 +46,7 @@ public struct ExpiringTodoRule: ConfigurationProviderRule, OptInRule {
 
     public func validate(file: SwiftLintFile) -> [StyleViolation] {
         // swiftlint:disable:next line_length
-        let regex = "\\b(?:TODO|FIXME)(?::|\\b)(?:.*)\\\(configuration.dateDelimiters.opening)(\\d{2,4}\\\(configuration.dateSeparator)\\d{2}\\\(configuration.dateSeparator)\\d{2,4})\\\(configuration.dateDelimiters.closing)"
+        let regex = "\\b(?:TODO|FIXME)(?::|\\b)(?:.*)\\\(configuration.dateDelimiters.opening)(\\d{1,4}\\\(configuration.dateSeparator)\\d{1,2}\\\(configuration.dateSeparator)\\d{1,4})\\\(configuration.dateDelimiters.closing)"
 
         return file.matchesAndSyntaxKinds(matching: regex).compactMap { checkingResult, syntaxKinds in
             guard