+++ CHANGELOG.md
@@ -26,6 +26,9 @@
   with a static member (e.g. `if number == .zero`).  
   [Marcelo Fabri](https://github.com/marcelofabri)
   [#3562](https://github.com/realm/SwiftLint/issues/3562)
+* Fix the regex for expiring-todos  
+  [namolnad](https://github.com/namolnad)
+  [#3597](https://github.com/realm/SwiftLint/pull/3597)
 
 ## 0.43.1: Laundroformat